Make AutoSource a template class
