+++ lib/fles_ipc/TimeslicePublisher.hpp
@@ -5,6 +5,10 @@
 
 #include "Sink.hpp"
 #include "StorableTimeslice.hpp"
+#include <algorithm>
+#include <boost/archive/binary_oarchive.hpp>
+#include <boost/iostreams/device/back_inserter.hpp>
+#include <boost/iostreams/stream.hpp>
 #include <string>
 #include <zmq.hpp>
 
@@ -14,18 +18,21 @@ namespace fles {
  * \brief The TimeslicePublisher class publishes serialized timeslice data sets
  * to a zeromq socket.
  */
-class TimeslicePublisher : public TimesliceSink {
+template <class Base, class Derived> class Publisher : public Sink<Base> {
 public:
   /// Construct timeslice publisher sending at given ZMQ address.
-  TimeslicePublisher(const std::string& address, uint32_t hwm = 1);
+  explicit Publisher(const std::string& address, uint32_t hwm = 1) {
+    publisher_.set(zmq::sockopt::sndhwm, int(hwm));
+    publisher_.bind(address.c_str());
+  }
 
   /// Delete copy constructor (non-copyable).
-  TimeslicePublisher(const TimeslicePublisher&) = delete;
+  Publisher(const Publisher&) = delete;
   /// Delete assignment operator (non-copyable).
-  void operator=(const TimeslicePublisher&) = delete;
+  void operator=(const Publisher&) = delete;
 
   /// Send a timeslice to all connected subscribers.
-  void put(std::shared_ptr<const fles::Timeslice> timeslice) override {
+  void put(std::shared_ptr<const Base> timeslice) override {
     do_put(*timeslice);
   };
 
@@ -34,7 +41,23 @@ private:
   zmq::socket_t publisher_{context_, ZMQ_PUB};
   std::string serial_str_;
 
-  void do_put(const fles::StorableTimeslice& timeslice);
+  void do_put(const Derived& timeslice) {
+    // serialize timeslice to string
+    serial_str_.clear();
+    boost::iostreams::back_insert_device<std::string> inserter(serial_str_);
+    boost::iostreams::stream<boost::iostreams::back_insert_device<std::string>>
+        s(inserter);
+    boost::archive::binary_oarchive oa(s);
+    oa << timeslice;
+    s.flush();
+
+    zmq::message_t message(serial_str_.size());
+    std::copy_n(static_cast<const char*>(serial_str_.data()), message.size(),
+                static_cast<char*>(message.data()));
+    publisher_.send(message, zmq::send_flags::none);
+  }
 };
+
+using TimeslicePublisher = Publisher<Timeslice, StorableTimeslice>;
 
 } // namespace fles
