+++ lib/fles_ipc/TimesliceSubscriber.hpp
@@ -3,8 +3,9 @@
 /// \brief Defines the fles::TimesliceSubscriber class.
 #pragma once
 
+#include "Source.hpp"
 #include "StorableTimeslice.hpp"
-#include "TimesliceSource.hpp"
+#include "Timeslice.hpp"
 #include <boost/archive/binary_iarchive.hpp>
 #include <boost/iostreams/device/array.hpp>
 #include <boost/iostreams/stream.hpp>
@@ -16,17 +17,21 @@ namespace fles {
  * \brief The TimesliceSubscriber class receives serialized timeslice data sets
  * from a zeromq socket.
  */
-class TimesliceSubscriber : public TimesliceSource {
+template <class Base, class Derived> class Subscriber : public Source<Base> {
 public:
   /// Construct timeslice subscriber receiving from given ZMQ address.
-  explicit TimesliceSubscriber(const std::string& address, uint32_t hwm);
+  Subscriber(const std::string& address, uint32_t hwm) {
+    subscriber_.set(zmq::sockopt::rcvhwm, int(hwm));
+    subscriber_.connect(address.c_str());
+    subscriber_.set(zmq::sockopt::subscribe, "");
+  }
 
   /// Delete copy constructor (non-copyable).
-  TimesliceSubscriber(const TimesliceSubscriber&) = delete;
+  Subscriber(const Subscriber&) = delete;
   /// Delete assignment operator (non-copyable).
-  void operator=(const TimesliceSubscriber&) = delete;
+  void operator=(const Subscriber&) = delete;
 
-  ~TimesliceSubscriber() override = default;
+  ~Subscriber() override = default;
 
   /**
    * \brief Retrieve the next item.
@@ -35,19 +40,43 @@ public:
    *
    * \return pointer to the item, or nullptr if end-of-file
    */
-  std::unique_ptr<StorableTimeslice> get() {
-    return std::unique_ptr<StorableTimeslice>(do_get());
-  };
+  std::unique_ptr<Derived> get() { return std::unique_ptr<Derived>(do_get()); };
 
   [[nodiscard]] bool eos() const override { return eos_flag; }
 
 private:
-  StorableTimeslice* do_get() override;
+  Derived* do_get() override {
+    if (eos_flag) {
+      return nullptr;
+    }
+
+    zmq::message_t message;
+    [[maybe_unused]] auto result = subscriber_.recv(message);
+
+    boost::iostreams::basic_array_source<char> device(
+        static_cast<char*>(message.data()), message.size());
+    boost::iostreams::stream<boost::iostreams::basic_array_source<char>> s(
+        device);
+    boost::archive::binary_iarchive ia(s);
+
+    Derived* sts = nullptr;
+    try {
+      sts = new Derived(); // NOLINT
+      ia >> *sts;
+    } catch (boost::archive::archive_exception& e) {
+      delete sts; // NOLINT
+      eos_flag = true;
+      return nullptr;
+    }
+    return sts;
+  }
 
   zmq::context_t context_{1};
   zmq::socket_t subscriber_{context_, ZMQ_SUB};
 
   bool eos_flag = false;
 };
+
+using TimesliceSubscriber = Subscriber<Timeslice, StorableTimeslice>;
 
 } // namespace fles
