+++ lib/fles_ipc/TimesliceView.hpp
@@ -11,6 +11,8 @@
 
 namespace fles {
 
+template <class Base, class Derived> class Receiver;
+
 /**
  * \brief The TimesliceView class provides access to the data of a single
  * timeslice in memory.
@@ -25,7 +27,7 @@ public:
   ~TimesliceView() override = default;
 
 private:
-  friend class TimesliceReceiver;
+  friend class Receiver<Timeslice, TimesliceView>;
   friend class StorableTimeslice;
 
   TimesliceView(
